refactor(nlp): vocabulary extraction moved to Dataset class
