+++ packages/botonic-nlp/src/dataset/dataset.ts
@@ -1,4 +1,7 @@
-import { shuffle as shuffleArray } from '../utils/array-utils'
+import { Preprocessor } from '../preprocess'
+import { PADDING_TOKEN } from '../preprocess/constants'
+import { Vocabulary } from '../preprocess/vocabulary'
+import { flatten, shuffle as shuffleArray, unique } from '../utils/array-utils'
 import { DatasetLoader } from './loader'
 import { Sample } from './types'
 
@@ -27,6 +30,16 @@ export class Dataset {
       trainSet: new Dataset(this.classes, this.entities, trainSamples),
       testSet: new Dataset(this.classes, this.entities, testSamples),
     }
+  }
+
+  extractVocabulary(preprocessor: Preprocessor): Vocabulary {
+    const sequences = this.samples.map(sample =>
+      preprocessor.preprocess(sample.text, PADDING_TOKEN)
+    )
+    const tokens = flatten(sequences)
+    const uniqueTokens = unique(tokens)
+    const filteredTokens = uniqueTokens.filter(token => token !== PADDING_TOKEN)
+    return new Vocabulary(filteredTokens)
   }
 
   get length(): number {
